Praticamente tudo terminado, agora falta a escala de bradem e base de dados
