+++ backend/middleware/check-auth.js
@@ -4,7 +4,7 @@ module.exports = (req, res, next) => {
     
     try {
         const token = req.headers.authorization.split(" ")[1];
-        const decodedToken = jwt.verify(token, "senhamuitograndealgoritmo");
+        const decodedToken = jwt.verify(token, process.env.JWT_KEY);
         req.dadosUsuario = {email: decodedToken.email, usuarioId: decodedToken.usuarioId};
         next();
     }catch(error) {
