+++ backend/routes/fichas.js
@@ -1,135 +1,19 @@
 const express = require ("express");
 
-const Ficha = require('../models/ficha');
 const checkAuth = require("../middleware/check-auth"); //middleware para autenticar o usuário, se não estiver autenticado não pode criar, editar ou apagar ficha alguma
 
 const router = express.Router();
 
-router.post("", checkAuth, (req, res, next) => {
-    const ficha = new Ficha({
-        nome: req.body.nome,
-        matricula: req.body.matricula,
-        leito: req.body.leito,
-        data: req.body.data,
-        percepSens: req.body.percepSens,
-        umidade: req.body.umidade,
-        atividade: req.body.umidade,
-        mobilidade: req.body.mobilidade,
-        nutricao: req.body.nutricao,
-        fricscisal: req.body.fricscisal,
-        score: req.body.score,
-        criador: req.dadosUsuario.usuarioId
-    });
-    ficha.save().then(fichacreated => {
-        res.status(201).json({
-            message: 'Ficha salva com sucesso!',
-            fichaId: fichacreated._id
-        });
-    })
-    .catch(error => {
-        res.status(500).json({
-            message: 'Falha no envio da ficha!'
-        });
-    });
-});
+const FichaController = require("../controllers/fichas");
 
-router.put("/:id", checkAuth, (req, res, next) => {
-    const ficha = new Ficha({
-        _id: req.body.id,
-        nome: req.body.nome,
-        matricula: req.body.matricula,
-        leito: req.body.leito,
-        data: req.body.data,
-        percepSens: req.body.percepSens,
-        umidade: req.body.umidade,
-        atividade: req.body.umidade,
-        mobilidade: req.body.mobilidade,
-        nutricao: req.body.nutricao,
-        fricscisal: req.body.fricscisal,
-        score: req.body.score
-    });
-    Ficha.updateOne({_id: req.params.id}, ficha).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Ficha editada com sucesso!" });
-    })
-    .catch(error => {
-        res.status(500).json({
-            message: 'Falha na edição da ficha!'
-        });
-    });
-});
+router.post("", checkAuth, FichaController.createFicha);
 
-//5ce72a5361cba61abc8383d4
+router.put("/:id", checkAuth, FichaController.editFicha);
 
-router.get("", checkAuth, (req, res, next) => {
-    if (req.dadosUsuario.usuarioId === '5ce72a5361cba61abc8383d4') {
-        Ficha.find().then(documents => {
-            res.status(200).json({
-            message: 'fichas carregadas com sucesso!',
-            fichas: documents
-            });
-        });
-    }else {
-        Ficha.find({criador: req.dadosUsuario.usuarioId}).then(documents => {
-            res.status(200).json({
-            message: 'fichas carregadas com sucesso!',
-            fichas: documents
-            });
-        })
-        .catch(error => {
-            res.status(500).json({
-                message: 'Falha ao carregar as fichas'
-            });
-        });
-    }
-});
+router.get("", checkAuth, FichaController.getFichas);
 
-// router.get("", (req, res, next) => {
-//     const pageSize = +req.query.pagesize;
-//     const paginaAtual = +req.query.page;
-//     const fichaQuery = Ficha.find();
-//     // const fichaQuery = Ficha.find({criador: req.dadosUsuario.usuarioId});
-//     let fichasLoad;
-//     if (pageSize && paginaAtual) {
-//         fichaQuery
-//         .skip(pageSize * (paginaAtual - 1))
-//         .limit(pageSize);
-//     }
+router.get("/:id", FichaController.getFicha);
 
-//     fichaQuery.then(documents => {
-//         res.status(200).json({
-//             message: 'fichas carregadas com sucesso!',
-//             fichas: documents
-//         });
-//     });
-// });
-
-router.get("/:id", (req, res, next) => {
-    Ficha.findById(req.params.id).then(ficha => {
-        if(ficha) {
-            res.status(200).json(ficha);
-        } else {
-            res.status(404).json({message: 'Ficha não encontrada!'})
-        }
-    })
-    .catch(error => {
-        res.status(500).json({
-            message: 'Falha ao carregar a ficha'
-        });
-    });
-});
-
-router.delete("/:id", checkAuth, (req, res, next) => {
-    Ficha.deleteOne({_id: req.params.id}).then(
-        result => {
-        //console.log(result);
-        res.status(200).json({ message: "Ficha deletada" });
-    })
-    .catch(error => {
-        res.status(500).json({
-            message: 'Não foi possível excluir a ficha'
-        });
-    });;
-});
+router.delete("/:id", checkAuth, FichaController.deleteFicha);
 
 module.exports = router;
