--- ionic.config.json
@@ -1,5 +1,0 @@
-{
-  "name": "AppLesaoPressaoV1",
-  "integrations": {},
-  "type": "angular"
-}
