+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+    "env": {
+        "MONGO_ATLAS_PSW": "KnGVmJ7tRoWynMYQ",
+        "JWT_KEY": "senhamuitograndealgoritmo"
+    }
+}
